+++ app/src/main/java/io/github/tonnyl/sample/MainActivity.kt
@@ -2,18 +2,22 @@ package io.github.tonnyl.sample
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import io.github.tonnyl.sample.databinding.ActivityMainBinding
 import io.github.tonnyl.whatsnew.WhatsNew
 import io.github.tonnyl.whatsnew.item.WhatsNewItem
 import io.github.tonnyl.whatsnew.util.PresentationOption
-import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
+
+    private val binding by lazy(LazyThreadSafetyMode.NONE) {
+        ActivityMainBinding.inflate(layoutInflater)
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        textView.setOnClickListener {
+        binding.textView.setOnClickListener {
             WhatsNew.newInstance(
                     WhatsNewItem("Nice Icons", "Completely customize colors, texts and icons.", R.drawable.ic_heart),
                     WhatsNewItem("Such Easy", "Setting this up only takes 2 lines of code, impressive you say?", R.drawable.ic_thumb_up),
